// 23729131_LeThiKimNgan_Tuan4/src/types/user.ts
/**
 * Định nghĩa Type/Interface cho đối tượng User từ https://jsonplaceholder.typicode.com/users/1
 */

export interface Geo {
  lat: string;
  lng: string;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo?: Geo;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address?: Address;
  phone: string;
  website: string;
  company?: Company;
}